C4-1-Routes Added for Show & Add Tasks
